Implemented new command handler decorator to simplify writing commands.
As an example, I removed the old !8ball, !elXX, and !election if blocks
in on_privmsg and on_pubmsg. I added @command_handlers with regexes that
match the old manual parsing of these commands.

This removes some of the redundancy of the original implementation, as
90% of the code between the 10 shorthand election blocks was shared. We
should see a similar improvement for the other shorthand commands.

Even for simple commands like !8ball, I believe this method of specifying
how commands are parsed is a little more explicit and easier to maintain,
but I look forward to comments on my method.